+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **Breaking!** - Removed the read-only `wrappedElement` property from the `<Tether>` and `<BaseTetherToolip>` component in favor of attachments.
 - Added functionality to allow _multi-child tethering_. Previously, it was only required to have exactly one child element in a `<Tether>`.
 - Added the _`tether(...)` attachment_, which can be used via `{@attach tether(...)}` on elements/components directly.
+- Added the optional `measureAnchor` prop to `<Tether>`, which can be specified to conditionally enable/disable the non-trivial computation of the position and size of wrapped elements.
+- Added the optional `measureAnchorWhile(...)` predicate prop to `<BaseTetherTooltip>`. By default, tooltips now only calculate layout while their anchor is focused or hovered.
 
 ### Preview 0.3.0 (2026-06-11)